+++ backend/src/main/java/shop/ninescent/cartItem/controller/CartController.java
@@ -1,13 +1,15 @@
-package shop.ninescent.mall.cartItem.controller;
+package shop.ninescent.cartItem.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
-import shop.ninescent.mall.cartItem.domain.CartItem;
-import shop.ninescent.mall.cartItem.domain.CartItemDTO;
-import shop.ninescent.mall.cartItem.service.CartItemService;
-import shop.ninescent.mall.cartItem.service.CartService;
+import shop.ninescent.cartItem.domain.Cart;
+import shop.ninescent.cartItem.domain.CartItem;
+import shop.ninescent.cartItem.domain.CartItemDTO;
+import shop.ninescent.cartItem.service.CartItemService;
+import shop.ninescent.cartItem.service.CartService;
 
 import java.util.List;
