+++ backend/src/main/java/shop/ninescent/cartItem/domain/Cart.java
@@ -1,4 +1,4 @@
-package shop.ninescent.mall.cartItem.domain;
+package shop.ninescent.cartItem.domain;
 
 import jakarta.persistence.*;
 import lombok.Data;
