+++ backend/src/main/java/shop/ninescent/cartItem/repository/CartRepository.java
@@ -1,8 +1,9 @@
-package shop.ninescent.mall.cartItem.repository;
+package shop.ninescent.cartItem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shop.ninescent.mall.cartItem.domain.Cart;
+import shop.ninescent.cartItem.domain.Cart;
+import shop.ninescent.cartItem.domain.CartItem;
 
 import java.util.List;
