+++ backend/src/main/java/shop/ninescent/cartItem/service/CartItemService.java
@@ -1,12 +1,12 @@
-package shop.ninescent.mall.cartItem.service;
+package shop.ninescent.cartItem.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import shop.ninescent.mall.cartItem.domain.Cart;
-import shop.ninescent.mall.cartItem.domain.CartItem;
-import shop.ninescent.mall.cartItem.repository.CartItemRepository;
-import shop.ninescent.mall.cartItem.repository.CartRepository;
+import shop.ninescent.cartItem.domain.Cart;
+import shop.ninescent.cartItem.domain.CartItem;
+import shop.ninescent.cartItem.repository.CartItemRepository;
+import shop.ninescent.cartItem.repository.CartRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
